+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project (hashcat.launcher) will be documented in thi
 
 ### Changed:
  - Moved default extra arguments to general (#66)
+ - Aborted status color to yellow (#71)
 
 ## [1.0.0] 22/12/2021
 ### Added: